kernal_webapp/requirements.txt
appdirs==1.4.4
cachelib==0.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
distlib==0.3.0
filelock==3.0.12
Flask==1.1.2
Flask-Session==0.3.2
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2-binary==2.8.5
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.17
urllib3==1.25.9
virtualenv==20.0.20
Werkzeug==1.0.1
zipp==3.1.0
spotipy==2.16.1
